\name{otolith-defunct}
\alias{routine2}
\alias{checkart}
\alias{checkOrder}
\alias{orderOutline}
\title{Otolith defunct functions}
\description{
The following functions are defunct and not used anymore in the current release}
\usage{
routine2(...)
checkart(...)
checkOrder(...)
orderOutline(...)
}
\arguments{
  \item{...}{defunct arguments}
}