\name{readcurvi}
\alias{readcurvi}
\title{Read curvilinear semi-landmarks file}
\usage{
readcurvi(rfolder, write = FALSE, scale = NULL)
}
\arguments{
  \item{rfolder}{Path of folder containing the individual
  semi-landmark files (the 'curvillinearLand_' files)}

  \item{write}{logical. whether to write the array of
  landkmark configurations into a \code{.tps} file}

  \item{scale}{scale in the unit of um/px, i.e. 1 pixel of
  the image equal to how many um?}
}
\value{
array of semi-lamdmark configurations
}
\description{
a wrapper function to read the individual semi-landmark
files generated by \code{\link{img2landmark}} function, an
alternative from reading the \code{.tps} file generated
from \code{img2landmark}.
}
\seealso{
\code{\link{img2landmark}}
}

