\name{otolith-package}
\alias{otolith-package}
\alias{otolith}
\docType{package}
\title{
Identification of fish otolith using digital images
}
\description{
A package provides tools to process digital images of fish otoliths and building identification tools based on geometric morphometrics.
}
\details{
\tabular{ll}{
Package: \tab otolith\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2014-04-26\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Wong Jin Yung <wongjinyung@gmail.com>
}
\keyword{ package, otolith, geometric morphometrics }