\name{otolith-package}
\alias{otolith-package}
\alias{otolith}
\docType{package}
\title{
Identification of fish otolith using digital images
}
\description{
A package provides tools to process digital images of fish otoliths and building identification tools based on geometric morphometrics.
}
\details{
This package form a part of the project on the study of Malaysian fish otolith with the following members:
  \itemize{
    \item Wong Jin Yung \email{wongjinyung@gmail.com}
    \item Loh Kar Hoe \email{khloh@um.edu.my}
    \item Cecilia Chu \email{coccynx@yahoo.com}
    \item Chong Ving Ching \email{chong@um.edu.my}
  }
}